- Moved logic to get hero details in local data source. - Add logic to save rating to hero.

+++ app/src/main/java/com/nyan/twtest/domain/entity/HeroEntity.kt
@@ -4,5 +4,5 @@ data class HeroEntity(
     val id: Int,
     val name: String?,
     val photo: Int?,
-    val rating: Int = 0,
+    val rating: Float = 0f,
 )
